feat(xlang): expand field tag IDs to signed int32 (#3982)
## Why?

Draft: contributor-authored purpose, design rationale, and tradeoff
discussion will be added before
this pull request is marked ready for maintainer review.

## What does this PR do?

- Defines the cross-language field-tag domain as `0 <= tag_id < 2^29`.
- Uses signed 32-bit field-tag carriers across all supported runtimes
while retaining `-1` only as
  the internal no-tag sentinel.
- Corrects and documents the compact/extended tag encoding, including
the independent long-field-name
  extension rule.
- Adds owner-local range and duplicate validation in annotations, code
generators, metadata codecs,
  and the Fory compiler.
- Keeps wire field identity separate from compatible-reader local
dispatch identity.
- Adds cross-language coverage for tag IDs `15`, `65551`, and
`536870911`.

## Related issues

None.

## Does this PR introduce any user-facing change?

- [x] Does this PR introduce any public API change?
- [x] Does this PR introduce any binary protocol compatibility change?

Field tags now accept the complete range from `0` through `536870911`.
Existing valid field tags and
their wire bytes remain unchanged; the new range permits larger
identifiers through the existing
extended encoding.

## Benchmark

Paired latency deltas below compare this branch with frozen
`apache/main` at
`26b6b6cefca63aaab0195523ce817d8b3da2845b`. Negative values are
improvements.

| Runtime/path | Serialize | Deserialize |
| --- | ---: | ---: |
| C++ | +0.070% | +0.858% |
| Java runtime codegen | -0.014% | -2.383% |
| Java static generated | +0.779% | -0.457% |
| Rust | -3.740% | -0.281% |
| Go | -0.424% | +0.504% |
| Python Cython | -2.151% | +0.184% |
| Python pure | -1.185% | -1.648% |
| JavaScript | -0.416% | -0.075% |
| C# final review-fix boundary | -0.813% | -1.094% |
| Dart | -0.365% | +0.351% |
| Swift | -3.315% | +0.255% |

Kotlin and Scala changes are compile-time validation changes and use the
measured Java
static-generated runtime path. All reported paired medians are below the
`+1.0%` regression ceiling.

+++ cpp/fory/meta/field_info.h
@@ -193,7 +193,7 @@ struct FieldNodeSpec {
 };
 
 struct FieldMeta {
-  int16_t id_ = -1;
+  int32_t id_ = -1;
   bool has_id_ = false;
   bool nullable_ = false;
   bool ref_ = false;
@@ -203,7 +203,7 @@ struct FieldMeta {
   bool compress_ = true;
   int16_t type_id_override_ = -1;
 
-  constexpr FieldMeta id(int16_t v) const {
+  constexpr FieldMeta id(int32_t v) const {
     auto copy = *this;
     copy.id_ = v;
     copy.has_id_ = true;
@@ -258,7 +258,7 @@ struct FieldMeta {
 };
 
 constexpr FieldMeta F() { return FieldMeta{}; }
-constexpr FieldMeta F(int16_t id) { return FieldMeta{}.id(id); }
+constexpr FieldMeta F(int32_t id) { return FieldMeta{}.id(id); }
 
 namespace T {
